kind: run via sudo when using rootless podman
Using podman (on Fedora 36) the daemon/daemonset container fails to
start with the following error:

  Message:   mount: /sys/fs/bpf: permission denied.

and the pod goes into an endless crashloopbackoff state.

If you switch to docker then this problem doesn't occur so I suspect
this is just a rootless podman issue (i.e., lack of privileges)
because if the scripts to create/delete the KIND cluster are invoked
using sudo then the error doesn't occur.

I don't know whether we want to encourage or use sudo to start the
KIND cluster but invoking `kind create cluster` via sudo allows the
daemon container to start. The known working e2e tests also run
through to completion.

This patch always runs via sudo. If we wanted to support rootless
podman then this would need to switch on whether podman or docker is
in use.

Signed-off-by: Andrew McDermott <amcdermo@redhat.com>